[JM-1301] Now using BIGINT instead of INTEGER for postgres entryStamp.
git-svn-id: http://svn.igniterealtime.org/svn/repos/openfire/branches/openfire_3_5_0@10088 b35dd754-fafc-0310-a699-88a17e54d16e

+++ src/database/upgrade/16/openfire_postgresql.sql
@@ -2,7 +2,7 @@
 CREATE TABLE jiveSecurityAuditLog (
   msgID                 INTEGER         NOT NULL,
   username              VARCHAR(64)     NOT NULL,
-  entryStamp            INTEGER         NOT NULL,
+  entryStamp            BIGINT          NOT NULL,
   summary               VARCHAR(255)    NOT NULL,
   node                  VARCHAR(255)    NOT NULL,
   details               TEXT,
